Fix MSVC build error (#19913)
* Explicitly cast double to int then cast int to enum
* Write a line of comment

+++ src/item.cpp
@@ -431,7 +431,8 @@ bool item::is_sided() const {
 }
 
 side item::get_side() const {
-    return static_cast<side>( get_var( "lateral", static_cast<int>( side::BOTH ) ) );
+    // MSVC complains if directly cast double to enum
+    return static_cast<side>( static_cast<int>( get_var( "lateral", static_cast<int>( side::BOTH ) ) ) );
 }
 
 bool item::set_side (side s) {
